apps/backend/requirements/development.txt
-r base.txt

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.12.1
pylint==3.0.3
pylint-django==2.5.5

# Debugging
django-debug-toolbar==4.2.0
ipython==8.20.0
